+++ src/basic/meson.build
@@ -44,8 +44,9 @@ basic_sources = files(
         'initrd-util.c',
         'inotify-util.c',
         'io-util.c',
-        'iovec-util.c',
         'ioprio-util.c',
+        'iovec-util.c',
+        'iovec-wrapper.c',
         'label.c',
         'limits-util.c',
         'locale-util.c',
