+++ src/test/test-log.c
@@ -6,6 +6,7 @@
 #include "format-util.h"
 #include "io-util.h"
 #include "iovec-util.h"
+#include "iovec-wrapper.h"
 #include "log.h"
 #include "process-util.h"
 #include "string-util.h"
